# docs/openapi.yaml
openapi: 3.0.3
info:
  title: lensing API
  version: 0.1.0
  description: >-
    HTTP API of the lensing prediction lab: build feature datasets from a
    Qdrant corpus, launch and monitor training runs, promote runs into named
    models and predict against them, maintain the curated best-models group,
    manage model definitions (predictor + hyperparameter presets), author
    manual entries, and blend the test-split predictions of finished runs.
    Domain-specific vocabulary (entity noun, target noun, field names) comes
    from domain.toml — see GET /api/domain.
servers:
  - url: /api

# The API is unauthenticated (single-user lab server).
security: []
...
paths:
...
  /models:
    get:
      tags: [models]
      summary: List promoted models
      responses:
        '200':
          description: Model records, newest first.
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items: { $ref: '#/components/schemas/ModelRecord' }
    post:
      tags: [models]
      summary: Promote a run into a model
      description: >-
        Copy a finished run's artifacts into a named model. Remote-worker runs
        are materialized from the database first. Triggers a best-models
        recompute.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, run_id]
              properties:
                name: { type: string, description: "Model name; must match ^[a-z0-9][a-z0-9-]{0,63}$." }
                run_id: { type: string }
                notes: { type: string, nullable: true }
      responses:
        '201':
          description: Model created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModelRecord' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /models/{name}:
    get:
      tags: [models]
      summary: Get a model
      description: Model record, contract summary, and training hyperparams.
      parameters: [{ $ref: '#/components/parameters/ModelName' }]
      responses:
        '200':
          description: Model detail.
          content:
            application/json:
              schema:
                type: object
                properties:
                  record: { $ref: '#/components/schemas/ModelRecord' }
                  contract: { $ref: '#/components/schemas/ContractSummary' }
                  hyperparams: { nullable: true, type: object, additionalProperties: true }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [models]
      summary: Delete a model
      description: >-
        Triggers a best-models recompute: a deleted group member is backfilled
        by the next-best candidate (to keep a top-ranked run OUT of the group,
        exclude it via PUT /best-models instead).
      parameters: [{ $ref: '#/components/parameters/ModelName' }]
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
        '404':
          $ref: '#/components/responses/NotFound'
  /models/{name}/viz:
    get:
      tags: [models]
      summary: Get model architecture diagram
      parameters: [{ $ref: '#/components/parameters/ModelName' }]
      responses:
        '200':
          description: SVG diagram.
          content:
            image/svg+xml:
              schema: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
...
  /models/{name}/contract:
    get:
      tags: [models]
      summary: Get model inference contract
      description: "What callers must send: required fields, embedding dim, target transform."
      parameters: [{ $ref: '#/components/parameters/ModelName' }]
      responses:
        '200':
          description: Contract summary.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContractSummary' }
        '404':
          $ref: '#/components/responses/NotFound'
...
  /models/{name}/predict:
    post:
      tags: [models]
      summary: Predict against a model
      description: >-
        One-shot predict: featurizes raw items and/or Qdrant point ids under
        the model's frozen contract and returns target-space predictions.
      parameters: [{ $ref: '#/components/parameters/ModelName' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PredictRequest' }
      responses:
        '200':
          description: Predictions.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PredictResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /best-models:
    get:
      tags: [best-models]
      summary: Get the best-models group
      description: >-
        The auto-maintained top-N-by-primary-metric group of promoted models,
        with curation state (pins and excludes).
      responses:
        '200':
          description: Best-models group.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BestModelGroup' }
    put:
      tags: [best-models]
      summary: Curate the best-models group
      description: >-
        Apply curation deltas: pins force-include promoted models regardless
        of rank; excludes (run ids or model names) bar candidates from
        auto-selection. A recompute follows; the updated group is returned.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CurateRequest' }
      responses:
        '200':
          description: Updated group.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BestModelGroup' }
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /best-models/predict:
    post:
      tags: [best-models]
      summary: Predict with every group member
      description: >-
        Fan out the predict request to every group member and aggregate a
        median consensus per input row. A failing member becomes a warning;
        only an empty group or all members failing is an error.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PredictRequest' }
      responses:
        '200':
          description: Per-member predictions plus consensus.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GroupPredictResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
...
    ModelName:
      name: name
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflicting state (e.g. name taken, run still running).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    InferencePrediction:
      type: object
      description: One inference prediction (target space; no ground truth).
      properties:
        row_id: { type: integer, format: int64 }
        predicted: { type: number }
      required: [row_id, predicted]
...
    TargetInfo:
      type: object
      properties:
        field: { type: string }
        transform: { type: string, enum: [log1p, none] }
      required: [field, transform]
    FeatureConfig:
      type: object
      description: >-
        Feature configuration a dataset was built with. `fields` is the
        generic, authoritative per-field enable map; the named flags
        (bedrooms, property_type, neighborhood_top_n, city, province,
        cluster) are legacy wire names kept for pre-domain artifacts and
        clients, mirrored from the effective enables.
      properties:
        pca_dims: { type: integer }
        fields:
          type: object
          additionalProperties: { type: boolean }
          description: Per-field enables, keyed by domain field name or toggle-group name.
        vocab_top_n:
          type: object
          additionalProperties: { type: integer }
          description: Per-categorical vocabulary-size overrides (field name → top-N).
        coordinate_bounds:
          type: array
          nullable: true
          description: "[[lat_min, lat_max], [lon_min, lon_max]] frozen into the artifact."
          items: { type: array, items: { type: number }, minItems: 2, maxItems: 2 }
          minItems: 2
          maxItems: 2
        bedrooms: { type: boolean, description: Legacy flag. }
        property_type: { type: boolean, description: Legacy flag. }
        neighborhood_top_n: { type: integer, description: Legacy flag; 0 disables. }
        city: { type: boolean, description: Legacy flag. }
        province: { type: boolean, description: Legacy flag. }
        cluster: { type: boolean, description: Legacy flag. }
        raw_numerics: { type: boolean, default: false }
        area_content_backfill: { type: boolean, default: false }
        coordinates: { type: boolean, default: false }
        impute_numerics: { type: boolean, default: false }
      required: [pca_dims]
...
    ModelRecord:
      type: object
      description: A promoted run.
      properties:
        name: { type: string }
        run_id: { type: string }
        predictor: { type: string }
        dataset_id: { type: string }
        created_at: { type: string, description: RFC 3339 UTC. }
        notes: { type: string, nullable: true }
      required: [name, run_id, predictor, dataset_id, created_at]
    ContractSummary:
      type: object
      description: Caller-relevant slice of a model's frozen featurization contract.
      properties:
        contract_version: { type: integer }
        n_cols: { type: integer }
        input_fields:
          type: object
          properties:
            required_numeric: { type: array, items: { type: string } }
            required_categorical: { type: array, items: { type: string } }
            embedding_dim: { type: integer }
        target: { $ref: '#/components/schemas/TargetInfo' }
        feature_config: { $ref: '#/components/schemas/FeatureConfig' }
        pca_dims: { type: integer }
    BestModelEntry:
      type: object
      description: One member of the best-models group, with ranking provenance.
      properties:
        name: { type: string, description: Promoted model name. }
        rank: { type: integer, description: 1 = best. }
        metric: { type: string, description: Primary metric label the ranking used. }
        metric_value: { type: number, description: "Primary metric value, target space." }
        run_id: { type: string }
        predictor: { type: string }
        dataset_id: { type: string }
        source:
          type: string
          enum: [auto, pinned]
          description: auto = selected by the metric ranking; pinned = curated in.
        selected_at: { type: string, description: RFC 3339 UTC. }
      required: [name, rank, metric, metric_value, run_id, predictor, dataset_id, source, selected_at]
    BestModelGroup:
      type: object
      description: >-
        The best-models group document; rewritten wholesale on each recompute.
      properties:
        primary_metric: { type: string }
        size: { type: integer, description: Configured group-size cap (pins may exceed it). }
        entries:
          type: array
          items: { $ref: '#/components/schemas/BestModelEntry' }
        pinned: { type: array, items: { type: string }, description: Model names force-included by curation. }
        excluded: { type: array, items: { type: string }, description: Run ids / model names force-excluded by curation. }
        updated_at: { type: string, description: RFC 3339 UTC; last recompute. }
      required: [primary_metric, size, entries, updated_at]
    CurateRequest:
      type: object
      description: Curation deltas for the best-models group. All fields optional.
      properties:
        pin: { type: array, items: { type: string }, description: Model names to force-include. }
        exclude: { type: array, items: { type: string }, description: Run ids or model names to bar from auto-selection. }
        unpin: { type: array, items: { type: string } }
        unexclude: { type: array, items: { type: string } }
    MemberPredictions:
      type: object
      properties:
        name: { type: string }
        rank: { type: integer }
        predictions:
          type: array
          items: { $ref: '#/components/schemas/InferencePrediction' }
      required: [name, rank, predictions]
    ConsensusPoint:
      type: object
      description: Median across members for one input row.
      properties:
        row_id: { type: integer, format: int64 }
        predicted: { type: number }
        n_models: { type: integer, description: Members that produced a prediction for this row. }
      required: [row_id, predicted, n_models]
    GroupPredictResponse:
      type: object
      properties:
        members:
          type: array
          items: { $ref: '#/components/schemas/MemberPredictions' }
        consensus:
          type: array
          items: { $ref: '#/components/schemas/ConsensusPoint' }
        warnings: { type: array, items: { type: string } }
      required: [members, consensus, warnings]
...
    PredictRequest:
      type: object
      description: Raw items and/or Qdrant point ids. At least one must be non-empty.
      properties:
        items:
          type: array
          items: { type: object, additionalProperties: true }
          description: Raw entries (content + metadata fields) to featurize.
        point_ids:
          type: array
          items: { type: integer, format: int64 }
          description: Qdrant point ids to fetch and featurize.
    PredictResponse:
      type: object
      properties:
        predictions:
          type: array
          items: { $ref: '#/components/schemas/InferencePrediction' }
        warnings: { type: array, items: { type: string } }
      required: [predictions, warnings]
...